requirements.txt
matplotlib==3.4.2
numpy
Pillow
scikit-image
pytest